pyproject.toml
[tool.poetry]
name = "drf-market"
version = "0.1.0"
description = ""
authors = ["iliaman <iliaman1993@gmail.com>"]
readme = "README.md"
packages = [{include = "drf_market"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.7"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
psycopg = "^3.1.13"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
